requirements.txt
streamlit
langchain
langchain-google-genai
chromadb
sentence-transformers
unstructured[md]
python-dotenv